requirements.txt
camel-ai[rec]
pydantic~=2.11.3
python-dotenv
python-dateutil~=2.9.0.post0
requests-oauthlib
pandas